Destruction of cache is made optional.

+++ package/R/compileWorkflows.R
@@ -4,6 +4,7 @@
 #'
 #' @param dir String containing the path to the directory of book chapters.
 #' Usually the same as \code{dir} in \code{\link{setupBookChapters}}.
+#' @param fresh Logical scalar indicating whether or not to delete old caches.
 #'
 #' @return
 #' All workflow chapters are (re)compiled to generate the corresponding \code{*_cache} directories.
@@ -13,16 +14,22 @@
 #' Compilation is performed in an isolated session using \code{\link{r}} from the \pkg{callr} package.
 #' This ensures that settings from one chapter do not affect the next chapter.
 #'
+#' The default \code{fresh=TRUE} is recommended for a rigorous recompilation,
+#' but setting it to \code{FALSE} may be more convenient for quick debugging cycles.
+#'
 #' @author Aaron Lun
 #'
 #' @export
 #' @importFrom callr r
 #' @importFrom rmarkdown render
-compileWorkflows <- function(dir=".") {
+compileWorkflows <- function(dir=".", fresh=TRUE) {
     all.files <- list.files(dir, pattern=sprintf("^P%i_W.*\\.Rmd", .workflow_part), full.names=TRUE)
     for (f in all.files) {
-        cache.loc <- sub("\\.Rmd$", "_cache", f)
-        unlink(cache.loc, recursive=TRUE)
+        if (fresh) {
+            cache.loc <- sub("\\.Rmd$", "_cache", f)
+            unlink(cache.loc, recursive=TRUE)
+        }
+
         r(function(input) { rmarkdown::render(input) }, args = list(input = f), 
             show=TRUE, spinner=FALSE)
     }
